requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
dnspython==2.4.2
mysql-connector-python==8.2.0
requests==2.31.0
